Move transformSnippetIndex to transformer file
Update tests.
Update references.
