+++ src/functions/transformers/index.js
@@ -6,10 +6,15 @@ import {
   stripExpertiseFromTags
 } from './transformTags';
 
+import {
+  transformSnippetIndex
+} from './transformSnippets';
+
 export {
   transformTagName,
   determineExpertiseFromTags,
   determineLanguageFromTags,
   stripLanguageFromTags,
-  stripExpertiseFromTags
+  stripExpertiseFromTags,
+  transformSnippetIndex
 };
